First Commit:
This is an example of a genetic algorithm applied to a regression
problem. We instantiate a polynomial and attempt to use GA methods
(crossover & mutation) to converge on the target function.

The coefficients are constrained to be in (-10,10) and the points used
to compute the RMSE (error) between the target and our best individual
are equally spaced in (-20,20).

Parameters include the population size, the degree of the polynomial and
the probability of mutation. Crossover involves some children produced
from averaging the best individuals and some produced from alternating
entries from parents.

Extensions could include:
- regression for arbitrary functions (non-polynomials)
- multi-variable regression (harder to visualise!)
- varying the points generated, and constraints on the coefficients
- using derivatives to help approximate
- exploring other crossover methods